Move Existing Requests to respective namespace
