+++ app/Http/Controllers/AccountSettings/AccountSettingsController.php
@@ -6,9 +6,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\UpdateUsernameRequest;
-use App\Http\Requests\UpdateEmailRequest;
-use App\Http\Requests\UpdatePasswordRequest;
+use App\Http\Requests\AccountSettings\UpdateUsernameRequest;
+use App\Http\Requests\AccountSettings\UpdateEmailRequest;
+use App\Http\Requests\AccountSettings\UpdatePasswordRequest;
 use App\Models\User;
 
 class AccountSettingsController extends Controller
